backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
timescaledb==0.2.0

# Redis and caching
redis==5.0.1
hiredis==2.2.3

# Message Queue (Kafka/RabbitMQ)
kafka-python==2.0.2
pika==1.3.2  # RabbitMQ client

# WebSocket
websockets==12.0
python-socketio==5.10.0

# HTTP clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data processing
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Machine Learning
tensorflow==2.15.0
scikit-learn==1.3.2

# Environment and config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Type checking
mypy==1.7.1
types-requests==2.31.0.10

